requirements.txt
adal==1.2.1
asn1crypto==0.24.0
azure-common==1.1.18
azure-mgmt-compute==4.4.0
backcall==0.1.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
clin28tools==0.1.6
colibricore==2.4.5
cryptography==2.5
decorator==4.3.2
dynaconf==1.2.0
httplib2==0.10.3
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.13.2
Keras==2.1.2
msrest==0.6.4
msrestazure==0.6.0
oauthlib==3.0.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
PyNLPl==1.2.5
python-box==3.2.4
python-dateutil==2.8.0
python-dotenv==0.10.1
rdflib==4.2.2
recurrentshop==1.0.0
requests==2.21.0
requests-oauthlib==1.2.0
seq2seq==1.0.0
six==1.12.0
toml==0.10.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
